BB - first working version of subinlets (for lack of a better term)
Option for subinlets are in main options in interface, under "channel".
Allows user to transfer water between channel branches (specify possible transfer discharge), or simply add or substract discharge from a channel cell, according to a discharge timeseries.

How to use: Create a map (subinlet.map) with id's. For every cell with a value above 0, Lisem will look for subinlet_<id>.txt, with <id> the absolute id value. If a specific value is present both positive and negative, they are linked, and water is transfered. If there is only the positive value, water is simply added or substracted according to the discharge timeseries.

Format for timeseries (units are in minutes, and cubic meters per second):

#Header
2                      #nr of columns
time                   #dexcription of columns
q
0     5                #data
1     6
3     8